feat: Implementação de Regra de negócio

+++ src/main/java/com/global/orbit_bridge/service/OrganizacaoService.java
@@ -3,7 +3,9 @@ package com.global.orbit_bridge.service;
 import com.global.orbit_bridge.dto.OrganizacaoDto;
 import com.global.orbit_bridge.model.Organizacao;
 import com.global.orbit_bridge.repository.OrganizacaoRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,17 +15,25 @@ public class OrganizacaoService {
     @Autowired
     private OrganizacaoRepository organizacaoRepository;
 
-    public void salvarOrganizacao(OrganizacaoDto organizacaoDto) {
+    @Transactional
+    public ResponseEntity<Organizacao> salvarOrganizacao(OrganizacaoDto organizacaoDto) {
         Organizacao organizacao = Organizacao.builder()
                 .nome(organizacaoDto.getNome())
                 .descricao(organizacaoDto.getDescricao())
                 .build();
 
         organizacaoRepository.save(organizacao);
+        return ResponseEntity.ok(organizacao);
     }
 
-    public void excluirOrganizacao(Long idOrganizacao) {
+    @Transactional
+    public ResponseEntity<Organizacao> excluirOrganizacao(Long idOrganizacao) {
+        if (!organizacaoRepository.existsById(idOrganizacao)){
+            return ResponseEntity.badRequest().build();
+        }
+
         organizacaoRepository.deleteById(idOrganizacao);
+        return ResponseEntity.noContent().build();
     }
 
     public List<Organizacao> listarOrganizacoes() {
